Moving Compare tests to the units they belong to
... as is the case for all CombinedUnits
